format blade, wyswietlanie errora dla rejestracji

+++ app/Http/Controllers/AuthController.php
@@ -1,10 +1,12 @@
 <?php
 
 namespace App\Http\Controllers;
+
 use Illuminate\Support\Str;
 use App\Models\User;
 use Carbon\Carbon;
 use Illuminate\Contracts\Session\Session as SessionSession;
+use Illuminate\Support\Facades\Validator;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
@@ -13,6 +15,7 @@ use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Mail;
 use Illuminate\Support\Facades\Password;
 use Illuminate\Support\Facades\Session as FacadesSession;
+use Illuminate\Support\Facades\URL;
 use PgSql\Lob;
 
 class AuthController extends Controller
@@ -45,18 +48,19 @@ class AuthController extends Controller
         if (Auth::attempt($credentials)) {
             return (redirect()->intended(route('homepage')));
         }
-        return redirect(route('login'))->with("error", "login failed");
+        return redirect(route('login'))->with("error", "Logowanie nieudane");
     }
     function registerPost(Request $request)
     {
-        $request->validate([
+        $validator = Validator::make($request->all(), [
             'name' => [
-                'required'
+                'required',
+                'min:3',
+                'max:20'
             ],
             'email' => [
                 'required',
-                'email',
-                'unique:users,email'
+                'email'
             ],
             'password' => [
                 'required',
@@ -67,10 +71,12 @@ class AuthController extends Controller
                 'digits:9'
             ]
         ]);
+        if ($validator->fails()) {
+            return redirect()->back()->with('error', 'Niepoprawne dane');
+        }
         $userExists = User::where('email', $request->email)->exists();
-        if($userExists)
-        {
-            return redirect(route('login'))->with("error", "Podany email istnieje już w bazie");
+        if ($userExists) {
+            return redirect(route('login'))->with('error', "Podany email istnieje już w bazie");
         }
         $user = new User();
         $user->name = $request->name;
@@ -82,8 +88,8 @@ class AuthController extends Controller
 
         if ($user->save() && Auth::attempt($credentials)) {
             return redirect()->intended(route('homepage'))->with("success", "user created succesfully");
-        }else
-        return redirect(route('register'))->with("error", "Failed to create user");
+        } else
+            return redirect(route('register'))->with("error", "Failed to create user");
     }
     public function logout()
     {
@@ -108,26 +114,25 @@ class AuthController extends Controller
         ]);
         $token = Str::random(64);
         DB::table('password_resets')->insert([
-            'email'=> $request->email,
-            'token'=>$token,
-            'created_at'=> Carbon::now()
+            'email' => $request->email,
+            'token' => $token,
+            'created_at' => Carbon::now()
         ]);
 
-        Mail::send('emails.forgot-password', ['token'=>$token], function ($message) use ($request){
-           $message->to($request->email);
-           $message->subject("Reset Password"); 
-        }); 
+        Mail::send('emails.forgot-password', ['token' => $token], function ($message) use ($request) {
+            $message->to($request->email);
+            $message->subject("Reset Password");
+        });
         return redirect()->to(route('password.request'));
-
     }
-    public function resetPassword($token)
+    public function resetPassword($token, Request $request)
     {
-        return view('new-password', compact('token'));
+        return view('new-password', compact('token'), ['request' => $request]);
     }
     public function resetPasswordPost(Request $request)
     {
-       
-        $request->validate([
+
+        $validator = Validator::make($request->all(), [
             'token' => [
                 'required'
             ],
@@ -141,7 +146,7 @@ class AuthController extends Controller
                 'string',
                 'min:7',
                 'confirmed'
-            ],'password_confirmation' => [
+            ], 'password_confirmation' => [
                 'required',
                 'string',
                 'min:7',
@@ -149,15 +154,30 @@ class AuthController extends Controller
             ]
 
         ]);
-        $updatePassword = DB::table('password_resets')->where([
-            'email' =>$request->email,
-            'token' =>$request->token
-        ])->first();
-        if(!$updatePassword){
-            return redirect()->to(route('password.reset'))->with("error", "invalid");
+        if ($validator->fails()) {
+            return redirect()->back()->with('error', 'Niepoprawne dane');
         }
-        User::where('email', $request->email)->update(['password'=>Hash::make($request->password)]);
-        DB::table('password_resets')->where(['email'=>$request->email])->delete();
+        $updatePassword = DB::table('password_resets')->where([
+            'email' => $request->email,
+            'token' => $request->token,
+        ])->first();
+        if ($updatePassword === null || $updatePassword->created_at === null) {
+            return redirect()->back()->with('error', 'Niepoprawny e-mail');
+        }
+        $createdat = ($updatePassword->created_at);
+        $now = Carbon::now();
+        $diff = $now->diffInMinutes($createdat);
+        if ($request->email != $updatePassword->email) {
+            return redirect()->route('password.reset')->with('error', 'Niepoprawny e-mail');
+        }
+        if ($updatePassword === null || $updatePassword->token === null || $diff > 3) {
+            return redirect()->route('password.request')->with('error', 'link wygasł');
+        }
+        if (!$updatePassword) {
+            return redirect()->route('password.reset')->with("error", "invalid");
+        }
+        User::where('email', $request->email)->update(['password' => Hash::make($request->password)]);
+        DB::table('password_resets')->where(['email' => $request->email])->delete();
         return redirect()->to(route('login'))->with("success", "haslo zostało zmienione");
     }
 }
